pyproject.toml
[project]
name = "gmail-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "dataclasses>=0.8",
    "datetime>=5.5",
    "dotenv>=0.9.9",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-pubsub>=2.31.1",
    "mcp[cli]>=1.13.1",
    "openai>=1.101.0",
    "requests>=2.32.5",
    "response>=0.5.0",
    "supabase>=2.18.1",
]
